fix: build failures and clippy warnings across workspace
- Synchronize workspace versions to 0.0.1-canary and fix internal dependency mismatches.
- Improve scripts/set-version.sh to safely update only the first version occurrence.
- Fix mouse_absolute signature in InputInjector trait (i32 -> f32) across all platforms.
- Create wavry-common::helpers for shared security functions (constant_time_eq).
- Refactor MacScreenEncoder and FFI run_client to resolve too_many_arguments clippy warnings.
- Resolve WebRTC bridge naming collision and missing prost dependency in wavry-server.
- Fix undefined selected_codec in monitor selection handler.
- Update wavry-client CLI to match updated internal API signatures.

+++ crates/wavry-platform/src/lib.rs
@@ -37,7 +37,7 @@ impl InputInjector for UnsupportedInjector {
         bail!("input injection is not implemented for this platform")
     }
 
-    fn mouse_absolute(&mut self, _x: i32, _y: i32) -> Result<()> {
+    fn mouse_absolute(&mut self, _x: f32, _y: f32) -> Result<()> {
         bail!("input injection is not implemented for this platform")
     }
 }
